:bug: Force static render Admin routes, refactor consent-banner
