+++ aksel.nav.no/website/app/(routes)/(admin)/admin-dev/[[...index]]/page.tsx
@@ -2,6 +2,8 @@ import { AdminStudio } from "@/app/(routes)/(admin)/AdminStudio";
 
 export { metadata, viewport } from "next-sanity/studio";
 
+export const dynamic = "force-static";
+
 export default function StudioPage() {
   return <AdminStudio />;
 }
