+++ aksel.nav.no/website/package.json
@@ -80,6 +80,7 @@
     "swr": "^2.3.2",
     "tailwindcss": "^3.3.3",
     "tinycolor2": "^1.6.0",
+    "typescript-cookie": "^1.0.6",
     "zod": "^3.22.4"
   },
   "_note": "When upgrading @axe-core/playwright, try removing the resolution entry in the root package.json",
